requirements.txt
wheel
aiohttp[speedups]
aiofiles
aria2p
async_generator
beautifulsoup4>=4.8.1
cowpy>=1.1.0
psutil
wget
telegraph
TgCrypto>=1.2.2
cffi>=1.0.0
dnspython
emoji
PyPDF2
cryptg
patool
gitpython
asyncurban>=0.3.3
google-api-python-client==2.9.0
google-auth-httplib2
google-auth-oauthlib
googletrans==4.0.0-rc1
google_images_download
gTTS
gTTS-token
hachoir
REPO_LINK
heroku3>=3.4.0
httplib2
humanize
meval>=2.5
lxml
lyrics_extractor
oauth2client
Pillow>=8.1.0
fontTools
psycopg2
psycopg2-binary
pybase64>=0.4.0
pylast
pySmartDL
python-barcode
python-dotenv
pytz
qrcode
requests>=2.25.1
search-engine-parser>=0.6.1
speedtest-cli>=2.1.2
sqlalchemy<=1.3.23
telethon>=1.14.0
telethon-tgcrypto
telethon-session-sqlalchemy
urbandict>=0.5
wikipedia>=1.4.0
coffeehouse>=2.1.0
gitpython
git+https://github.com/TomyPrs/deezloader
git+https://github.com/johnwmillr/LyricsGenius.git
git+https://github.com/Ncode2014/search-engine-parser
lyricsgenius
PyDrive2
pyDownload
beautifulsoup4>=4.8.1
PyGithub
bs4
covid>=2.2.9
base
pip>=19.2.3
setuptools>=28.8.0
selenium>=3.141.0
python-dateutil
pyfiglet
jotquote
html_telegraph_poster
csvfaker
git+https://github.com/LonamiWebs/Telethon
git+https://github.com/ahupp/python-magic
pytube>=9.5.1
pymongo[srv]
youtube-search-python>=1.4.6
youtube-search>=2.1.0
youtube_dl>=2021.6.6
justwatch
zipfile38
validators
geopy
glitch_this
twitter_scraper
pyaztro
jikanpy
vcsi
lottie
wand
colour
scipy
wordcloud
opencv-python
pendulum
rarfile
natsort
hentai
Faker==8.1.2
js2py
aiofile
validators
pytest
